# src/app-server/infra/rest/generated/server/go/error_model.go
/*
 * Lab Resource Management
 *
 * This specification defines the APIs provided by the application to manage/share the lab resources by reserving and releasing it.
 *
 * API version: 1.0
 * Contact: mithun.bs@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ErrorModel struct {

	Message string `json:"message"`

	Code int32 `json:"code"`
}